+++ src/icp.ts
@@ -1,5 +1,6 @@
 import {ICPTs} from '../proto/ledger_pb';
 import {E8S_PER_ICP} from './constants/constants';
+import {FromICPStringError} from './types/errors';
 
 export class ICP {
   private constructor(private e8s: bigint) {}
@@ -15,14 +16,14 @@ export class ICP {
    * 1'234'567.8901
    * 1,234,567.8901
    */
-  public static fromString(amount: string): ICP | FromStringError {
+  public static fromString(amount: string): ICP | FromICPStringError {
     // Remove all instances of "," and "'".
     amount = amount.trim().replace(/[,']/g, "");
 
     // Verify that the string is of the format 1234.5678
     const regexMatch = amount.match(/\d*(\.\d*)?/);
     if (!regexMatch || regexMatch[0] !== amount) {
-      return FromStringError.INVALID_FORMAT;
+      return FromICPStringError.INVALID_FORMAT;
     }
 
     const [integral, fractional] = amount.split(".");
@@ -33,18 +34,18 @@ export class ICP {
       try {
         e8s += BigInt(integral) * E8S_PER_ICP;
       } catch {
-        return FromStringError.INVALID_FORMAT;
+        return FromICPStringError.INVALID_FORMAT;
       }
     }
 
     if (fractional) {
       if (fractional.length > 8) {
-        return FromStringError.FRACTIONAL_MORE_THAN_8_DECIMALS;
+        return FromICPStringError.FRACTIONAL_MORE_THAN_8_DECIMALS;
       }
       try {
         e8s += BigInt(fractional.padEnd(8, "0"));
       } catch {
-        return FromStringError.INVALID_FORMAT;
+        return FromICPStringError.INVALID_FORMAT;
       }
     }
 
@@ -62,7 +63,3 @@ export class ICP {
   }
 }
 
-export enum FromStringError {
-  FRACTIONAL_MORE_THAN_8_DECIMALS,
-  INVALID_FORMAT,
-}
